yarn buildのエラーに対応
